requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
rich>=13.7.0
aiohttp>=3.9.0
aiohttp-socks>=0.9.0
pydantic>=2.5.0
